chore: allocate script to prefab, Electricwire change
