--- Assets/SWPPT3/Prefabs/Puzzle/ElectricWire.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: dbc858894383b9d429bcf0186fe02c8b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
