+++ Assets/SWPPT3/Prefabs/Puzzle/Light_Line.prefab
@@ -10,6 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 8241524119736188774}
   - component: {fileID: -7298873488949143058}
+  - component: {fileID: 3914616279602080613}
   m_Layer: 0
   m_Name: Light_Line
   m_TagString: Untagged
@@ -55,6 +56,20 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
+--- !u!114 &3914616279602080613
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3678918952838657729}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 720894524c224583b5123c0737417b54, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  stateSources: []
+  animator: {fileID: -7298873488949143058}
 --- !u!1 &4527852408709611683
 GameObject:
   m_ObjectHideFlags: 0
