+++ Assets/SWPPT3/Prefabs/Puzzle/Metal_Cube.prefab
@@ -12,7 +12,8 @@ GameObject:
   - component: {fileID: 4709948878179989563}
   - component: {fileID: 7402209992574110961}
   - component: {fileID: 7478655169163834185}
-  - component: {fileID: -7141302848823913908}
+  - component: {fileID: 4047322057194843869}
+  - component: {fileID: -3986992414241368853}
   m_Layer: 0
   m_Name: Metal_Cube
   m_TagString: Untagged
@@ -99,19 +100,27 @@ MeshCollider:
   m_Convex: 0
   m_CookingOptions: 30
   m_Mesh: {fileID: -6835905690748749900, guid: 1ad5b5271fcab442890e8813f6fe1f04, type: 3}
---- !u!54 &-7141302848823913908
-Rigidbody:
+--- !u!114 &4047322057194843869
+MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 9059292695080840484}
-  serializedVersion: 2
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 0
-  m_CollisionDetection: 0
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e551fbd8743c4f50b026883974fca6ff, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &-3986992414241368853
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9059292695080840484}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0c09447a90b2486dbaf0f9f60b29ab64, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
