+++ Assets/SWPPT3/Prefabs/Puzzle/Round_Room_Sphere_Fog.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 928411532287679304}
   - component: {fileID: 928411532287679307}
   - component: {fileID: 928411532287679306}
+  - component: {fileID: -1105901957426897061}
   m_Layer: 0
   m_Name: Round_Room_Sphere_Fog
   m_TagString: Untagged
@@ -97,3 +98,15 @@ SphereCollider:
   serializedVersion: 2
   m_Radius: 0.5
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &-1105901957426897061
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 928411532287679309}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e3d2d5816d2c4c72b698aeea2b09f27a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
