+++ Assets/SWPPT3/Prefabs/Puzzle/Shield.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 8056305843128070129}
   - component: {fileID: 8056305843128070130}
   - component: {fileID: -5186212655999948723}
+  - component: {fileID: 5516704682588829139}
   m_Layer: 0
   m_Name: Shield
   m_TagString: Untagged
@@ -120,3 +121,18 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
+--- !u!114 &5516704682588829139
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8056305843128070131}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a57b55a3ab994eb09ef4f58bcf483fe5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  stateSources: []
+  collider: {fileID: 8056305843128070130}
+  animator: {fileID: -5186212655999948723}
