+++ Assets/SWPPT3/Prefabs/Puzzle/Wood_Box.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 7621708895271697327}
   - component: {fileID: 5629860802133118821}
   - component: {fileID: 4977694923314212169}
-  - component: {fileID: 4051083254505867502}
+  - component: {fileID: -7873784755181315289}
   m_Layer: 0
   m_Name: Wood_Box
   m_TagString: Untagged
@@ -98,19 +98,15 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.74355185, y: 0.740143, z: 0.74365073}
   m_Center: {x: 0, y: 0.00001642108, z: -0.00000011920929}
---- !u!54 &4051083254505867502
-Rigidbody:
+--- !u!114 &-7873784755181315289
+MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 6134383050356027056}
-  serializedVersion: 2
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 0
-  m_CollisionDetection: 0
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 049246bd5088433cbf69b64e1b7309ea, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
