+++ Assets/SWPPT3/Scripts/Main/Prop/ElectricWire.cs
@@ -1,24 +1,7 @@
-using UnityEngine;
-
 namespace SWPPT3.Main.Prop
 {
-    public class ElectricWire : StateDst
+    public class ElectricWire : StateSource
     {
-        [SerializeField] private Animator animator;
 
-        protected override void OnSourceStateChanged(StateSource src, bool state)
-        {
-            if (animator != null)
-            {
-                if (this.State == On)
-                {
-                    animator.SetBool("IsOn", true);
-                }
-                else if (this.State == Off)
-                {
-                    animator.SetBool("IsOn", false);
-                }
-            }
-        }
     }
 }
