+++ Assets/SWPPT3/Scripts/Main/Prop/LightLine.cs
@@ -1,0 +1,24 @@
+using UnityEngine;
+
+namespace SWPPT3.Main.Prop
+{
+    public class LightLine : StateDst
+    {
+        [SerializeField] private Animator animator;
+
+        protected override void OnSourceStateChanged(StateSource src, bool state)
+        {
+            if (animator != null)
+            {
+                if (this.State == On)
+                {
+                    animator.SetBool("IsOn", true);
+                }
+                else if (this.State == Off)
+                {
+                    animator.SetBool("IsOn", false);
+                }
+            }
+        }
+    }
+}
